Fix: Move template implementation to header file to avoid linking issues
