+++ GenericList/GenericList.vcxproj
@@ -139,12 +139,13 @@
     </Link>
   </ItemDefinitionGroup>
   <ItemGroup>
-    <ClCompile Include="list.cpp" />
+    <ClCompile Include="list_test.cpp" />
     <ClCompile Include="master_file.cpp" />
     <ClCompile Include="string_list.cpp" />
   </ItemGroup>
   <ItemGroup>
     <ClInclude Include="list.h" />
+    <ClInclude Include="list_test.h" />
     <ClInclude Include="string_list.h" />
   </ItemGroup>
   <Import Project="$(VCTargetsPath)\Microsoft.Cpp.targets" />
